+++ src/ofxOscSubscriber.h
@@ -13,6 +13,8 @@
 #include "details/ofxpubsubosc_type_traits.h"
 
 namespace ofx {
+    using namespace ofxpubsubosc;
+    
     class OscSubscriberManager {
         struct AbstractParameter {
             virtual void read(ofxOscMessage &message) {}
@@ -99,7 +101,7 @@ namespace ofx {
             }
             
             template <typename U, size_t size>
-            inline void set(ofxOscMessage &m, U v[size], size_t offset = 0) {
+            inline void set(ofxOscMessage &m, U (&v)[size], size_t offset = 0) {
                 for(int i = 0; i < min(size, m.getNumArgs() / ofxpubsubosc::type_traits<U>::size); i++) {
                     set(m, v[i], offset + i * ofxpubsubosc::type_traits<U>::size);
                 }
@@ -130,7 +132,7 @@ namespace ofx {
             CallbackParameter(Callback callback)
             : callback(callback) {}
             
-            virtual void read(ofxOscMessage &message) {callback(message); }
+            virtual void read(ofxOscMessage &message) { callback(message); }
 
         private:
             Callback callback;
